Speeds up queries by checking field grammar at runtime
Often, the queries we make are not coming from a user's input and are
hardcoded by developers. Using a macro to build those queries has two
advantages. The first is correctness checking and developer joy. The
second is throughput. Third is a small memory decrease.

When grammars are turned into indices, they are parsed from a string
into a list. The parsing of the string can fail if the string is not
properly formatted. The sigil checks the string format at compile time
instead of runtime. That catches errors early and reduces developer
mistakes.

Throughput when using precompiled grammars was increased on multiple
platforms. I used Benchee on an M1 and M2 Mac. The M1 with a String was
1.29x slower. On the M2, the increase was even better, with the string
version being 3.65 times slower.

Memory consumption had a similar benefit. M1 Mac was 1.05x larger on M1
and 2.15x larger on the M2. I was surprised but ran the benchmarks
multiple times with similar results.

Further gains could be had by removing the ability to use Strings so
that no extra function heads need to be checked.