Adds compile time checking of HL7 paths
The advantages here are that compile time checking allows us to take the
hit for parsing only once and not when the system is running. It also
allows us to let developers know of mistakes before the system is
running.

I added a sigil_g so that in areas where strings aren't passed directly
to the parts functions we can use the sigil to do the compile time
check.

Deprecating the String versions will help eleminate the extra macros and
leave only the sigil and a function for building field grammars.

I created a struct so that users of the library can match on the struct
type in their own code. We have done that in another project so we can
tell if we have a path and behave differently than if a string is passed
into a certain location.