Fixes identifiers passed to HL7.Query macros
Until Strings are deprecated we need to handle them coming in as
identifiers. The issue is that the macro doesn't know the value at
compile time. We have to ensure that we fix up the variable contents
before using them. Once we remove the String support we can remove
this call.