Changes from macros to deprecations for HL7.Query
We originally went with using macros to rewrite HL7.Query functions.
Although this quickly moved us to using verified types instead of
`String` inputs it could be a lot of change to quickly. To facilitate a
slower transition we decided that rewriting with macros might not be the
way forward in a healthcare scenario.

To allow for testing and slower transitions we are opting for deprecated
functions instead.

HL7 does not have something that is called a Field Grammar in their
documentation, but what we called a Field Grammar they call a Path. To
keep our context in line with their documentation we are changing the
name of FieldGrammar to Path.

This change has a much lower impact and allows time to vet the changes
before jumping in head first.