Fix infinite recusion bug in HL7.Query.{find_first, update, find_all}
In the first iteration of the new query functions we
fell through to a clause that intended to new up an HL7.Query
in the case one was not provided (i.e., an HL7 struct or raw message
was instead), but unlike the main clauses did not match on an
HL7.Path struct.

However, the new functions will only work with a an HL7.Path{},
such that if a string or any other term is passed in, we'll cycle
endlessly between the set of two clauses.